refactor: extract BOM utilities to module
